Move dummies to separate file. Add pytest config to handle paths.
